requirements/base.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
cryptography==42.0.5
celery==5.3.6
redis==5.0.3
django-redis==5.4.0
drf-spectacular==0.27.1
django-ratelimit==4.1.0
python-dateutil==2.9.0
Pillow==10.2.0